+++ penpa-edit/prepare_deployment.js
@@ -23,7 +23,7 @@ if (ENABLE_DEPLOYMENT) {
 
   const user_script = document.createElement("script");
   user_script.type = "py";
-  user_script.src = "./main_pyscript.py";
+  user_script.src = "./main_deploy.py";
   user_script.async = false;
   user_script.setAttribute("config", "./pyscript.json");
   document.head.appendChild(user_script);
